fix(winui): cast ulong to int for CancellationTokenSource constructor
Math.Max returns ulong when mixing ulong and int operands.
CancellationTokenSource(TimeSpan) or (int ms) requires int.
Fixes CS1503: cannot convert from 'ulong' to 'int'.

+++ native/VaultPilot.WinUI/Views/SettingsDialog.xaml.cs
@@ -701,7 +701,7 @@ public sealed partial class SettingsDialog : ContentDialog
 
         try
         {
-            using var cts = new CancellationTokenSource(Math.Max(timeoutMs + 5_000, 10_000));
+            using var cts = new CancellationTokenSource((int)Math.Max(timeoutMs + 5_000, 10_000));
             var request = new ProviderConnectionRequest
             {
                 ApiBase = baseUrl,
